+++ src/main/java/com/example/iotl/controller/NewsController.java
@@ -1,14 +1,11 @@
 package com.example.iotl.controller;
 
 import java.util.List;
-import com.example.iotl.dto.NewsDto;
-import com.example.iotl.dto.NewsSearchResponse;
+import com.example.iotl.dto.news.NewsDto;
 import com.example.iotl.service.NewsService;
 import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
